Use System.getProperty(line.separator) instead of System.lineSeparator(). Refactor code
